后端日志系统，分页插件，dao自动生成器

+++ face-java/src/main/java/com/face/po/UserLoginPo.java
@@ -18,7 +18,7 @@ public class UserLoginPo implements Serializable {
 	/**
 	*主键
 	*/
-	private Integer userLoginId;
+	private Long userLoginId;
 
 	/**
 	*创建时间
